mid: 0000117 msg: Errors en la gestió d'enumeracions
git-svn-id: svn://svn.code.sf.net/p/gehelium/code/helium/branches/branch-2.1@211 e3271b2b-fd41-4d56-bedf-4b5636213925

+++ helium-webapp/src/main/java/net/conselldemallorca/helium/webapp/mvc/EnumeracioController.java
@@ -95,6 +95,7 @@ public class EnumeracioController extends BaseController {
 			if ("submit".equals(submit) || submit.length() == 0) {
 				command.setEntorn(entorn);
 				annotationValidator.validate(command, result);
+				new EnumeracioValidator(dissenyService).validate(command, result);
 		        if (result.hasErrors()) {
 		        	return "enumeracio/form";
 		        }
@@ -125,8 +126,13 @@ public class EnumeracioController extends BaseController {
 		Entorn entorn = getEntornActiu(request);
 		if (entorn != null) {
 			try {
-				dissenyService.deleteEnumeracio(id);
-				missatgeInfo(request, getMessage("info.enum.esborrat") );
+				Enumeracio enumeracio = dissenyService.getEnumeracioById(id);
+				if (enumeracio.getCamps().size() > 0) {
+					missatgeError(request, getMessage("error.enumeracio.emprada.camp") );
+				} else {
+					dissenyService.deleteEnumeracio(id);
+					missatgeInfo(request, getMessage("info.enum.esborrat") );
+				}
 			} catch (Exception ex) {
 	        	missatgeError(request, getMessage("error.esborrar.enum"), ex.getLocalizedMessage());
 	        	logger.error("No s'ha pogut esborrar l'enumeració", ex);
